BUG/MINOR: sample: fix backend direction flags consecutive to last fix
Commit 7a06ffb85 ("BUG/MEDIUM: sample: Cumulate frontend and backend
sample validity flags") introduced a typo confusing the request and
the response direction when checking for validity of a rule applied
to a backend. This was reported by Coverity in issue #1417.

This needs to be backported where the patch above is backported.

(cherry picked from commit b39e47a52b19e7b62727393f9b8e8bd36330b550)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>