MINOR: connection: implement function to update ALPN
Implement a new function to update the ALPN on an existing connection.
on an existing connection. The ALPN from the ssl context can be checked
to update the ALPN only if it is a subset of the context value.

This method will be useful to change a connection ALPN for websocket,
must notably if the server does not support h2 websocket through the
rfc8441 Extended Connect.

(cherry picked from commit 2454bda1405da260021547d1c9edd676b4219a22)
[ad: adjusted context because of inline function moved from connection.h
 to connection.c]
Signed-off-by: Amaury Denoyelle <adenoyelle@haproxy.com>