MINOR: memprof: report the delta between alloc and free on realloc()
realloc() calls are painful to analyse because they have two non-zero
columns and trying to spot a leaking one requires a bit of scripting.
Let's simply append the delta at the end of the line when alloc and
free are non-nul.

It would be useful to backport this to 2.4 to help with bug reports.

(cherry picked from commit 8cce4d79ff8e448b033aca9b4b09db3b7acf7928)
Signed-off-by: Willy Tarreau <w@1wt.eu>