SCRIPTS: git-show-backports: re-enable file-based filtering
The early version of the script used to support passing non-branch
arguments but as it evolved we lost that option. Let's use "--" as a
delimiter after the branch(es) to pass optional file names to filter
on. This is convenient to list missing patches on a specific set of
files.

(cherry picked from commit 0d026edaefcee30d5532a9fde7cbfbeed411379e)
Signed-off-by: Willy Tarreau <w@1wt.eu>