BUG/MINOR: cache: properly ignore unparsable max-age in quotes
When "max-age" or "s-maxage" receive their values in quotes, the pointer
to the integer to be parsed is advanced by one, but the error pointer
check doesn't consider this advanced offset, so it will not match a
parse error such as max-age="a" and will take the value zero instead.

This probably needs to be backported, though it's unsure it has any
effect in the real world.

(cherry picked from commit 1f38bdb3f63f3ce7422bedfe3a6e24c1619e9881)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>