Revert "BUG/MINOR: http-ana: Don't eval front after-response rules if stopped on back"
This reverts commit 597909f4e67866c4f3ecf77f95f2cd4556c0c638

http-after-response rules evaluation was changed to do the same that was
done for http-response, in the code. However, the opposite must be performed
instead. Only the rules of the current section must be stopped. Thus the
above commit is reverted and the http-response rules evaluation will be
fixed instead.

Note that only "allow" action is concerned. It is most probably an uncommon
action for an http-after-request rule.

This patch must be backported as far as 2.2 if the above commit was
backported.

(cherry picked from commit 4c5a591b1029ef03c3461a7771483aa523e3111a)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit f5db0fc6ccb5bdbbe0c6304968da180fae7d1d52)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>