DOC: lua: Be explicit with the Reply object limits
In HTTP, when a lua action is evaluated, a reply object can be used to send
a response to the client and interrupt the transaction. This reply object is
converted into HTX and is limited to the response channel buffer. Its size,
once converted, cannot exceed the buffer size. There is no streaming at this
stage. However, this limitation was not documented.

Note that, for now, there is no easy way to know if the reply will fit or
not int the response channel buffer. Thus the reply must be reasonably
small. Otherwise a 500-Internal-Error message is returned.

This patch is related to the issue #1447. It may be backported as far as
2.2.

(cherry picked from commit 7855b1973437926ee89be8c8e4620d627712cc81)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit 8cee915534b372adf75df1534b3bb5e743c39ab4)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit fd84357391d11221dcf2fceaa81c7a953d17ba8a)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>