BUG/MINOR: mworker: doesn't launch the program postparser
When in wait mode, the mworker-prog postparser is launched, but
unfortunately the child structure doesn't contain all required
information to be able to launch the test.

This test is only required when doing a configuration parsing.

Must be backported as far as 2.0.

(cherry picked from commit 5f47b2e28050066254397873401d3f087ecef1c1)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>