BUG/MEDIUM: mux-h1: Fix H1C_F_ST_SILENT_SHUT value
The commit 7700441a25 ("BUG/MINOR: mux-h1: Save shutdown mode if the
shutdown is delayed") introduced a regression during the backport process.
The H1C_F_ST_SILENT_SHUT value was not changed. However, on the 2.3 and
lowers, another H1C flag alreay have the value.

Because of this bug, H1C_F_ST_SILENT_SHUT and H1C_F_CS_SHUTDOWN shared the
same value. Mixed with some other bugs, this may lead to truncated messages.

This patch help to fix issue #1450. There is no upstream ID because the bug
only affects the 2.3 and lowers. The patch must be backported as far as 2.0.